+++ esp-hal/src/i2c/support/mod.rs
@@ -1,0 +1,3 @@
+pub mod eh;
+pub mod eh02;
+pub mod eh_a;
